Revert file change to fix test breakage in Travis.
