+++ android-sdk/src/main/java/com/optimizely/ab/android/sdk/OptimizelyManager.java
@@ -18,13 +18,17 @@ package com.optimizely.ab.android.sdk;
 
 import android.annotation.TargetApi;
 import android.app.Activity;
+import android.app.AlarmManager;
 import android.app.Application;
+import android.content.ComponentName;
 import android.content.Context;
 import android.content.Intent;
+import android.content.ServiceConnection;
 import android.content.res.Resources;
 import android.os.AsyncTask;
 import android.os.Build;
 import android.os.Bundle;
+import android.os.IBinder;
 import android.support.annotation.NonNull;
 import android.support.annotation.Nullable;
 import android.support.annotation.RawRes;
@@ -33,6 +37,9 @@ import android.support.annotation.VisibleForTesting;
 
 import com.optimizely.ab.Optimizely;
 import com.optimizely.ab.android.event_handler.OptlyEventHandler;
+import com.optimizely.ab.android.shared.Cache;
+import com.optimizely.ab.android.shared.Client;
+import com.optimizely.ab.android.shared.OptlyStorage;
 import com.optimizely.ab.android.shared.ServiceScheduler;
 import com.optimizely.ab.bucketing.UserExperimentRecord;
 import com.optimizely.ab.config.parser.ConfigParseException;
@@ -397,6 +404,86 @@ public class OptimizelyManager {
         @Override
         public void onActivityDestroyed(Activity activity) {
             // NO-OP
+        }
+    }
+
+    static class DataFileServiceConnection implements ServiceConnection {
+
+        @NonNull private final OptimizelyManager optimizelyManager;
+        private boolean bound = false;
+
+        DataFileServiceConnection(@NonNull OptimizelyManager optimizelyManager) {
+            this.optimizelyManager = optimizelyManager;
+        }
+
+        /**
+         * @hide
+         * @see ServiceConnection#onServiceConnected(ComponentName, IBinder)
+         */
+        @RequiresApi(api = Build.VERSION_CODES.HONEYCOMB)
+        @Override
+        public void onServiceConnected(ComponentName className,
+                                       IBinder service) {
+            // We've bound to DataFileService, cast the IBinder and get DataFileService instance
+            DataFileService.LocalBinder binder = (DataFileService.LocalBinder) service;
+            final DataFileService dataFileService = binder.getService();
+            if (dataFileService != null) {
+                DataFileClient dataFileClient = new DataFileClient(
+                        new Client(new OptlyStorage(dataFileService.getApplicationContext()), LoggerFactory.getLogger(OptlyStorage.class)),
+                        LoggerFactory.getLogger(DataFileClient.class));
+
+                DataFileCache dataFileCache = new DataFileCache(
+                        optimizelyManager.getProjectId(),
+                        new Cache(dataFileService.getApplicationContext(), LoggerFactory.getLogger(Cache.class)),
+                        LoggerFactory.getLogger(DataFileCache.class));
+
+                DataFileLoader dataFileLoader = new DataFileLoader(dataFileService,
+                        dataFileClient,
+                        dataFileCache,
+                        Executors.newSingleThreadExecutor(),
+                        LoggerFactory.getLogger(DataFileLoader.class));
+
+                dataFileService.getDataFile(optimizelyManager.getProjectId(), dataFileLoader, new DataFileLoadedListener() {
+                    @Override
+                    public void onDataFileLoaded(@Nullable String dataFile) {
+                        // App is being used, i.e. in the foreground
+                        AlarmManager alarmManager = (AlarmManager) dataFileService.getApplicationContext().getSystemService(Context.ALARM_SERVICE);
+                        ServiceScheduler.PendingIntentFactory pendingIntentFactory = new ServiceScheduler.PendingIntentFactory(dataFileService.getApplicationContext());
+                        ServiceScheduler serviceScheduler = new ServiceScheduler(alarmManager, pendingIntentFactory, LoggerFactory.getLogger(ServiceScheduler.class));
+                        if (dataFile != null) {
+                            AndroidUserExperimentRecord userExperimentRecord =
+                                    (AndroidUserExperimentRecord) AndroidUserExperimentRecord.newInstance(optimizelyManager.getProjectId(), dataFileService.getApplicationContext());
+                            optimizelyManager.injectOptimizely(dataFileService.getApplicationContext(), userExperimentRecord, serviceScheduler, dataFile);
+                        } else {
+                            // We should always call the callback even with the dummy
+                            // instances.  Devs might gate the rest of their app
+                            // based on the loading of Optimizely
+                            OptimizelyStartListener optimizelyStartListener = optimizelyManager.getOptimizelyStartListener();
+                            if (optimizelyStartListener != null) {
+                                optimizelyStartListener.onStart(optimizelyManager.getOptimizely());
+                            }
+                        }
+                    }
+                });
+            }
+            bound = true;
+        }
+
+        /**
+         * @hide
+         * @see ServiceConnection#onServiceDisconnected(ComponentName)
+         */
+        @Override
+        public void onServiceDisconnected(ComponentName arg0) {
+            bound = false;
+        }
+
+        boolean isBound() {
+            return bound;
+        }
+
+        void setBound(boolean bound) {
+            this.bound = bound;
         }
     }
